ErrorluTest

+++ e-commerce/src/main/java/com/example/e_commerce/repository/OrderRepository.java
@@ -1,9 +1,7 @@
 package com.example.e_commerce.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.e_commerce.entity.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
