+++ e-commerce/src/main/java/com/example/e_commerce/repository/ProductRepository.java
@@ -3,13 +3,14 @@ package com.example.e_commerce.repository;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.e_commerce.entity.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Integer>{
+public interface ProductRepository extends JpaRepository<Product, Long> {
 	List<Product> findAllByUserId(Integer id);
 	
 	List<Product> findAllByOrderByPriceAsc();
 
 	List<Product> findAllByOrderByPriceDesc();
+
 }
+
